finance/requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3
dateparser==0.7.0
Django==2.0.7
django-celery-beat==1.1.1
django-celery-results==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kombu==4.2.1
mysqlclient==1.3.13
Pillow==5.2.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-binance==0.6.9
python-dateutil==2.7.3
pytz==2018.5
regex==2018.7.11
requests==2.19.1
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
txaio==2.10.0
tzlocal==1.5.1
urllib3==1.23
vine==1.1.4
zope.interface==4.5.0
